feat: Add ERC20 inheritance to LiqidityManger contract
This commit adds the ERC20 inheritance to the LiqidityManger contract in order to enable token functionality. The contract now inherits from the ERC20 contract from the "@openzeppelin/contracts/token/ERC20" library. Additionally, the constructor has been updated to initialize the ERC20 token with the name "zkETH" and the symbol "zkETH". The _mint function is also called to mint tokens to the contract deployer when they deposit funds.